chore: add safeCreate/Update incubator

+++ src/components/IncubatorCreatePage/IncubatorCreate.tsx
@@ -5,7 +5,10 @@ import React from "react";
 import * as Sentry from "@sentry/nextjs";
 
 import { IncubatorForm } from "../IncubatorForm/IncubatorForm";
-import { createIncubator } from "@/app/api/incubators/actions/createIncubator";
+import {
+    createIncubator,
+    safeCreateIncubator,
+} from "@/app/api/incubators/actions/createIncubator";
 import { incubatorUpdateSchemaType } from "@/models/actions/incubator";
 import { Option } from "@/models/misc";
 import { saveImage } from "@/utils/file";
@@ -19,32 +22,35 @@ interface IncubatorInfoCreateProps {
 export const IncubatorCreate = (props: IncubatorInfoCreateProps) => {
     const save = async (data: incubatorUpdateSchemaType) => {
         try {
-            const newIncubator = await safeCreateIncubator({
+            const res = await safeCreateIncubator({
                 incubator: data,
             });
-            if (data.logo) {
-                saveImage({
-                    fileIdentifier: "logo",
-                    fileRelativeObjType: "incubator",
-                    fileObjIdentifier: newIncubator.ghid,
-                    file: data.logo,
-                });
-            }
+            if (res.success) {
+                if (data.logo) {
+                    saveImage({
+                        fileIdentifier: "logo",
+                        fileRelativeObjType: "incubator",
+                        fileObjIdentifier: res.data.ghid,
+                        file: data.logo,
+                    });
+                }
 
-            if (data.shouldDeleteLogo) {
-                await fetch("/api/image", {
-                    method: "DELETE",
-                    headers: {
-                        "Content-Type": "application/json",
-                    },
-                    body: JSON.stringify({
-                        fileObjIdentifier: newIncubator.ghid,
-                        fileIdentifer: "hero",
-                        fileRelativeObjType: "startup",
-                    }),
-                });
+                if (data.shouldDeleteLogo) {
+                    await fetch("/api/image", {
+                        method: "DELETE",
+                        headers: {
+                            "Content-Type": "application/json",
+                        },
+                        body: JSON.stringify({
+                            fileObjIdentifier: res.data.ghid,
+                            fileIdentifer: "hero",
+                            fileRelativeObjType: "startup",
+                        }),
+                    });
+                }
             }
             window.scrollTo({ top: 20, behavior: "smooth" });
+            return res;
         } catch (e) {
             window.scrollTo({ top: 20, behavior: "smooth" });
             Sentry.captureException(e);
